Initial commit: scaffold project
Creates a project scaffold via Bundler and sets the initial version to
0.0.0.

Specifically, this project scaffold was created by running the command
'bundle gem kramdown-syntax_tree_sitter --test=minitest --ci=github',
using version 2.3.22 of Bundler and an initial Git branch name of
'main'. After scaffolding the project, the version number was switched
from '0.1.0' to '0.0.0' so as to reserve version 0.1.0 for a future
release.

The choice of Minitest for the testing framework is based on the fact
that Kramdown uses it. A starter GitHub Actions CI workflow was elected
due to the decision to host this project on GitHub.