Rename starter GitHub Actions workflow
Renames the starter GitHub Actions workflow from 'Ruby' to 'Quality
Control'. The latter name is the name I generally use to indicate that a
workflow is used for verifying code quality, through the use of testing,
linting, and so on.

As of this commit, GitHub Actions is set up to run successfully (and,
moreover, in accordance with my personal preferences).