Patch issue with Bundler and project smoke test
Unfortunately, prior to this change, the built-in smoke test for this
project did not behave (and likely never behaved) correctly when used
alongside Bundler, which could lead to untrustworthy test results. This
patch appears to resolve that issue.

In particular, the intended approach for running the project's smoke
test is to run it as a Rake task. Project Rake tasks, in turn, should be
run from within a Bundler environment. Thus, the intended command for
running a smoke test is 'bundle exec rake smoke_test'. However, this
(in hindsight, obviously) means that the smoke test runs within the
context of a Bundler environment, which happens to implicitly include
the project as a gem. As a consequence, even though the smoke test is
intended specifically for testing a *new* install of the gem, when run
in this way, the smoke test can pass regardless of whether the gem is
installed at all. This can be verified by temporarily disabling the
dependency of the smoke_test task on the install task. Even though this
should cause the smoke_test task to fail, it does not.

The solution is to have the smoke_test task run the smoke test script
outside of the context of Bundler, as this commit codifies. With this
change, the smoke_test task now correctly fails if the gem has not been
installed outside of Bundler.

As part of this change, the Rake install and uninstall tasks were also
updated to play nicely with the smoke_test command.