Add Rust as a project dependency
Configures the project so that Rust is required as a dependency.
Although as of this commit Rust is technically not absolutely required,
it soon will be; hence this change prepares the project accordingly.
Specifically, some new Rake extensions tasks are added and included as
prerequisites for some existing tasks, an extension build script is
included (which uses the new Rake tasks under the hood) as part of the
gem installation pipeline, the Quality Control workflow to test
different versions of Rust, and the existing documentation is updated to
declare Rust as a requirement.