Use tree sitter_adapter package to highlight code
Adds a function that allows the tree_sitter_adapter library to escape
text, updates the main library method, updates the gem to invoke the
library in order to highlight code, and removes a test rendered obsolete
by these changes.

With this change, the Ruby and Rust components of the gem are fully
integrated. With this change, essentially the only MVP requirement
remaining  is building out the Tree-sitter adapter library completely.