Update plugin to depend on a parsers directory
Updates the plugin to depend on the existence of a Tree-sitter parsers
directory, which can be specified via the `tree_sitter_parsers_dir`
option. Also adds such a directory for the tests to depend on, updates
the tests to verify that the plugin requires the presence of the
directory, and updates the readme to describe this new behavior.