Patch/improve plugin parsers directory option
Updates the plugin so that user-set relative paths to the parsers
directory are now possible. Previously, setting the option
'tree_sitter_parsers_dir' to '~/tree_sitter_parsers' resulted in an
error, even though this is the directory used by default. With this
change, this is now possible.

Unfortunately, no tests were added to account for this change, as it
does not lend itself to testing very nicely.