Update plugin to ignore code without a language
Updates the plugin so that it no longer attempts to highlight code
blocks (and spans) that are missing a language. Also updates tests
accordingly.

Prior to this change, the plugin would crash if it encountered a code
block without a language. Now it effectively skips such blocks, and the
Kramdown highlighter interprets them as just plaintext.