Remove testing support for Rust version 1.64
Removes Rust version 1.64 from the Quality Control workflow as well as
from the list of officially supported versions in the documentation.
Also updates the changelog accordingly.

This Rust version is being removed from the list of supported versions
in preparation for bumping this project's Tree-sitter dependencies,
the higher versions of which only support Rust version 1.65 and
higher.