Remove testing support for Rust version 1.66.
Removes Rust version 1.66 from the Quality Control workflow as well as
from the list of officially supported versions in the documentation.
Also updates the changelog accordingly.

The decision to remove this Rust version from the list of officially
supported Rust versions is merely a practical consequence of the fact
that only so many versions of Rust can be tested by the Quality Control
workflow. For this project, the current (somewhat arbitrary) decision is
to test the latest four versions of Rust. Hence the introduction of Rust
1.70 also necessitates the removal of 1.66.

This is not to say that this version of Rust is truly no longer
supported; however, the decision to remove it from the list of
"officially" supported Rust versions is a consequence of an
unwillingness to claim support for a particular version without having
tests to justify this claim. Nevertheless, it is very likely that this
project will continue to work with this Rust version just as it did
before these changes were introduced.