Remove testing support for MacOS version 11
Removes MacOS 11 (Big Sur) from the Quality Control workflow and updates
the changelog accordingly.

The decision to remove this MacOS version from the list of officially
supported operating systems is both a practical consequence of the fact
that only so many operating systems can be tested by the Quality Control
workflow, and also a consequence of the fact that the 'macos-11' GitHub
Actions runner image has been deprecated and will no longer be available
after 2024-06-28.

This is not necessarily to say that this version of MacOS is truly no
longer supported, only that such support is no longer guaranteed now
that this tool is no longer tested on this operating system.