Remove testing support for Ruby version 2.7
Removes Ruby version 2.7 from the Quality Control workflow as well as
from the list of officially supported versions in the documentation.
Also raises the gem specification required ruby version to 3.0 and
updates the changelog accordingly.

The decision to remove this Ruby version from the list of officially
supported Ruby versions is mostly a practical consequence of the fact
that only so many versions of Ruby can be tested by the Quality Control
workflow. For this project, the current (somewhat arbitrary) decision is
to test the latest four versions of Ruby. Hence the introduction of Ruby
3.3 also necessitates the removal of Ruby 2.7. Moreover, Ruby 2.7 is in
end-of-life status and is no longer recommended for use.

This is not to say that this version of Ruby is truly no longer
supported; however, the decision to remove it from the list of
"officially" supported Ruby versions is a consequence of an
unwillingness to claim support for a particular version without having
tests to justify this claim. Nevertheless, it is likely that this
project will continue to work with this Ruby version just as it did
before these changes were introduced.